22
타일scene 렌더링오류 해결
렌더링 최적화
메뉴바 수정
gaper AI부여중

+++ 20220302/UI.cpp
@@ -90,7 +90,9 @@ void CUI::render(HDC _dc)
 		TransparentBlt(_dc
 			, static_cast<int>(vPos.x), static_cast<int>(vPos.y)
 			, static_cast<int>(vScale.x) * 2, static_cast<int>(vScale.y) * 2
-			, m_pTex->GetDC(), GetScale().x * m_vSlice.x, GetScale().y * m_vSlice.y
+			, m_pTex->GetDC()
+			, static_cast<int>(GetScale().x * m_vSlice.x)
+			, static_cast<int>(GetScale().y * m_vSlice.y)
 			, static_cast<int>(vScale.x) , static_cast<int>(vScale.y)
 			, RGB(255, 0, 255));
